pyproject.toml
[tool.poetry]
name = "geraimg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.4.0"

[tool.poetry.scripts ]
geraimg = "geraimg.geraimg:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
